Added daily printouts of failure classifications.
This is useful for seeing what the trend is on each type of failure. It is easy
to tell at a glance if a particular classification of failure is going away, or
has recently emerged with these being rendered.

Additionally, added a few more classifications. Now I do not see any
unclassified failures when I run against data downloaded today.